신청 상태에 따라 EventStatusTag를 반환하는 util 함수 구현
